requirements.txt
streamlit
langchain    
langchain-groq       
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
requests
tqdm
beautifulsoup4
transformers
